+++ egui_web/src/screen_reader.rs
@@ -4,6 +4,7 @@ pub struct ScreenReader {
 }
 
 #[cfg(not(feature = "screen_reader"))]
+#[allow(clippy::derivable_impls)] // False positive
 impl Default for ScreenReader {
     fn default() -> Self {
         Self {}